logique métier de reception et receptiondetails

+++ src/main/java/ma/baggar/bmsback/Dto/ReceptionDetailDto.java
@@ -6,11 +6,14 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import ma.baggar.bmsback.Entity.Article;
 import ma.baggar.bmsback.Entity.Reception;
+
+import java.math.BigDecimal;
+
 @Data @AllArgsConstructor @NoArgsConstructor
 public class ReceptionDetailDto {
 public Long id;
 private Article article;
 private float quantite;
-private float prixTtc;
+private BigDecimal prixTtc;
 private ReceptionDto reception;
 }
